+++ R/pirateplot_function.R
@@ -3,14 +3,15 @@
 #' The pirateplot function creates an RDI (Raw data, Descriptive and Inferential statistic) plot showing the relationship between 1 to 3 categorical independent variables and 1 continuous dependent variable.
 #'
 #' @param formula formula. A formula in the form \code{y ~ x1 + x2 + x3} indicating the vertical response variable (y) and up to three independent variables
-#' @param data dataframe. A dataframe containing the variables specified in formula.
+#' @param data Either a dataframe containing the variables specified in formula, a list of numeric vectors, or a numeric dataframe / matrix.
 #' @param plot logical. If \code{TRUE} (the default), thent the pirateplot is produced. If \code{FALSE}, the data summaries created in the plot are returned as a list.
 #' @param pal string. The color palette of the plot. Can be a single color, a vector of colors, or the name of a palette in the piratepal() function (e.g.; "basel", "google", "southpark"). To see all the palettes, run \code{piratepal(palette = "all", action = "show")}
+#' @param mix.col,mix.p Optional color mixing arguments to be passed to \code{piratepal}. See \code{?piratepal} for examples.
 #' @param point.col,bar.f.col,bean.b.col,bean.f.col,inf.f.col,inf.b.col,avg.line.col,bar.b.col,quant.col,point.bg string. Vectors of colors specifying the colors of the plotting elements. This will override values in the palette. f stands for filling, b stands for border.
 #' @param theme integer. An integer in the set 0, 1, 2 specifying a theme (that is, new default values for opacities and colors). \code{theme = 0} turns off all opacities which can then be individually specified individually.
 #' @param bar.f.o,point.o,inf.f.o,inf.b.o,avg.line.o,bean.b.o,bean.f.o,bar.b.o numeric. A number between 0 and 1 indicating how opaque to make the bars, points, inference band, average line, and beans respectively. These values override whatever is in the specified theme
 #' @param avg.line.fun function. A function that determines how average lines and bar heights are determined (default is mean).
-#' @param gl.col,back.col string. The color of the horizontal gridlines and plotting background.
+#' @param back.col string. Color of the plotting background.
 #' @param point.cex,point.pch,point.lwd numeric.  The size, pch type, and line width of raw data points.
 #' @param bean.lwd,bean.lty,inf.lwd,avg.line.lwd,bar.lwd numeric. Vectors of numbers customizing the look of beans and lines.
 #' @param width.min,width.max numeric. The minimum/maximum width of the beans.
@@ -23,15 +24,16 @@
 #' @param bw,adjust Arguments passed to density calculations for beans (see ?density)
 #' @param jitter.val numeric. Amount of jitter added to points horizontally. Defaults to 0.05.
 #' @param at integer. Locations of the beans. Especially helpful when adding beans to an existing plot with add = TRUE
-#' @param sortx string. How to sort the x values. Can be "sequential" (as they are found in the original dataframe), "alphabetical", or a string indicating a function (i.e.; "mean")
-#' @param add logical. Whether to add the pirateplot to an existing plotting space or not.
-#' @param evidence logical. Should Bayesian evidence be shown? (currently ignored)
+#' @param sortx string. How to sort the x values. Can be "sequential" (as they are found in the original dataframe), "alphabetical", or a string in the set ("mean", "median", "min", "max") indicating a function
+#' @param decreasing logical. If sortx is a named function, should values be sorted in decreasing order?
+#' @param add logical. Should the pirateplot elements be added to an existing plotting space?
 #' @param cap.beans logical. Should maximum and minimum values of the bean densities be capped at the limits found in the data? Default is FALSE.
 #' @param quant.length,quant.lwd numeric. Specifies line lengths/widths of \code{quant}.
 #' @param quant.boxplot logical. Should standard values be included?
 #' @param family a font family (Not currently in use)
-#' @param cex.lab,cex.axis Size of the labels and axes.
-#' @param gl.lwd,gl.lty Customization for grid lines.
+#' @param cex.lab,cex.axis,cex.names Size of the labels, axes, and bean names.
+#' @param gl numeric. Locations of the horizontal grid lines
+#' @param gl.lwd,gl.lty,gl.col Customization for grid lines. Can be entered as vectors for alternating gridline types
 #' @param bty,xlim,ylim,xlab,ylab,main,yaxt,xaxt General plotting arguments
 #' @param quant numeric. Adds horizontal lines representing custom quantiles.
 #' @param bar.b.lwd,line.fun,inf.o,bean.o,inf.col,theme.o,inf,inf.type,inf.band,bar.o,line.o,hdi.o depricated arguments
@@ -39,7 +41,7 @@
 #' @importFrom BayesFactor ttestBF
 #' @importFrom grDevices col2rgb gray rgb
 #' @importFrom graphics abline axis layout mtext par plot points polygon rasterImage rect segments text
-#' @importFrom stats density model.frame optimize rnorm t.test qbeta sd quantile IQR
+#' @importFrom stats density model.frame optimize rnorm t.test qbeta sd quantile IQR aggregate as.formula
 #' @importFrom utils vignette
 #' @export
 #' @examples
@@ -124,6 +126,8 @@ pirateplot <- function(
   plot = TRUE,
   avg.line.fun = mean,
   pal = "basel",
+  mix.col = "white",
+  mix.p = 0,
   back.col = NULL,
   point.cex = NULL,
   point.pch = NULL,
@@ -159,15 +163,16 @@ pirateplot <- function(
   adjust = 1,
   add = FALSE,
   sortx = "alphabetical",
+  decreasing = FALSE,
   cex.lab = 1,
   cex.axis = 1,
+  cex.names = 1,
   quant = NULL,
   quant.length = NULL,
   quant.lwd = NULL,
   quant.boxplot = FALSE,
   bty = "o",
-  evidence = FALSE,
-  cap.beans = FALSE,
+  cap.beans = TRUE,
   family = NULL,
   inf.method = "hdi",
   inf.within = NULL,
@@ -185,6 +190,7 @@ pirateplot <- function(
   main = NULL,
   yaxt = NULL,
   xaxt = NULL,
+  gl = NULL,
   gl.lwd = NULL,
   gl.lty = NULL,
   bar.b.lwd = NULL,
@@ -203,9 +209,9 @@ pirateplot <- function(
 #
 #
 #
-#
-  # formula = weight ~ Time
-  # data = ChickWeight
+# #
+  # formula = len ~ dose + supp
+  # data = ToothGrowth
   # plot = TRUE
   # avg.line.fun = mean
   # pal = "basel"
@@ -216,6 +222,7 @@ pirateplot <- function(
   # jitter.val = .03
   # theme = 1
   # bean.b.o = NULL
+  # quant.boxplot = FALSE
   # bean.f.o = NULL
   # point.o = NULL
   # bar.f.o = NULL
@@ -243,7 +250,8 @@ pirateplot <- function(
   # bw = "nrd0"
   # adjust = 1
   # add = FALSE
-  # sortx = "alphabetical"
+  # sortx = "mean"
+  # decreasing = TRUE
   # cex.lab = 1
   # cex.axis = 1
   # quant = NULL
@@ -273,16 +281,22 @@ pirateplot <- function(
   # line.fun = NULL
   # inf.o = NULL
   # bean.o = NULL
+  # bar.o = NULL
+  # line.o = NULL
   # inf.col = NULL
   # theme.o = NULL
   # inf = NULL
   # inf.type = NULL
   # inf.band = NULL
+  # cap.beans = TRUE
+  # #
   #
-  # formula = weight ~ Time
-  # data = ChickWeight
-  # theme = 2
-  # main = "theme = 2"
+  #
+  #
+  # formula = NULL
+  # data = list(rnorm(100),rnorm(20))
+  # ylab = ""
+  # sortx = "mean"
 
 # -----
 #  SETUP
@@ -376,8 +390,7 @@ if(is.null(bar.o) == FALSE) {
 
 # Look for missing critical inputs
 {
-if(is.null(data)) {stop("You must specify a dataframe in the data argument!")}
-if(is.null(formula) | class(formula) != "formula") {stop("You must specify a valid formula in the formula argument!")}
+if(is.null(data)) {stop("You must specify data in the data argument")}
 }
 
 # Set some defaults
@@ -394,6 +407,56 @@ if(inf.method %in% c("sd", "se") & is.null(inf.p)) {
 
 }
 
+
+# If no formula, than reshape data
+if(is.null(formula)) {
+
+  if(class(data) %in% c("data.frame", "matrix")) {
+
+    # data <- data.frame(a = rnorm(100), b = rnorm(100), c = rnorm(100))
+    # data <- matrix(rnorm(100), nrow = 20, ncol = 5)
+    #
+
+    data <- as.data.frame(data)
+    iv.levels <- names(data)
+
+    data <- stats::reshape(data, direction = "long", varying = list(1:ncol(data)))
+    data <- data[,1:2]
+    names(data) <- c("group", "y")
+
+    for(i in 1:length(iv.levels)) {
+
+      data$group[data$y == i] <- iv.levels[i]
+
+    }
+
+
+    formula <- y ~ group
+
+  }
+
+  if(class(data) == "list") {
+
+    if(is.null(names(data))) {names(data) <- paste0("V", 1:length(data))}
+
+    iv.levels <- names(data)
+
+    # Convert list to dataframe
+
+    data.df <- do.call("rbind", lapply(1:length(data), FUN = function(x) {
+
+      data.frame(group = rep(iv.levels[x], length(data[[x]])),
+                 y = unlist(data[[x]]))
+
+    }))
+
+    data <- data.df
+    formula <- y ~ group
+
+  }
+
+
+}
 
 # Reshape dataframe to include relevant variables
 {
@@ -452,6 +515,11 @@ data.i <- data.i[,1:min(ncol(data.i), 3)]
 
 # Determine levels of each IV
 {
+
+  if((substr(sortx, 1, 1) %in% c("a", "s") |
+     sortx %in% c("mean", "median", "min", "max")) == FALSE
+     ) {stop("sortx argument is invalid. use 'alphabetical', 'sequential', 'mean', 'median', 'min', or 'max'")}
+
 if(substr(sortx, 1, 1) == "a") {
 
   iv.levels <- lapply(2:ncol(data.i),
@@ -465,9 +533,44 @@ if(substr(sortx, 1, 1) == "s") {
 
 }
 
+if(sortx %in% c("mean", "median", "min", "max")) {
+
+  agg <- aggregate(formula, data = data.i, FUN = get(sortx))
+
+
+  if(ncol(agg) == 2) {
+
+    agg <- agg[order(agg[,2], decreasing = decreasing),]
+    iv.levels <- list(paste(agg[,1]))
+
+  }
+
+  if(ncol(agg) == 3) {
+
+  agg.1 <- aggregate(as.formula(paste(dv.name, "~", all.iv.names[1])),
+                     data = data.i, FUN = get(sortx))
+
+  agg.1 <- agg.1[order(agg.1[,2], decreasing = decreasing),]
+  iv.levels <- list(paste(agg.1[,1]))
+
+  agg.2 <- aggregate(as.formula(paste(dv.name, "~", all.iv.names[2])),
+                     data = data.i, FUN = get(sortx))
+
+  agg.2 <- agg.2[order(agg.2[,2], decreasing = decreasing),]
+  iv.levels[2] <- list(paste(agg.2[,1]))
+
+
+    }
+
+
+
+}
+
+
 iv.lengths <- sapply(1:length(iv.levels), FUN = function(x) {length(iv.levels[[x]])})
 iv.names <- names(data.i)[2:ncol(data.i)]
 subplot.n.iv <- length(iv.levels)
+
 }
 
 # Set up bean info
@@ -526,6 +629,8 @@ for(bean.i in 1:n.beans) {
   summary$avg[bean.i] <- avg.line.fun(dv.i)
 
   # Calculate inference
+
+  if(length(dv.i) > 0) {
 
   # Binary data.i
 
@@ -648,6 +753,16 @@ for(bean.i in 1:n.beans) {
   summary$inf.lb[bean.i] <- inf.lb
   summary$inf.ub[bean.i] <- inf.ub
 
+  }
+
+  if(length(dv.i) == 0) {
+
+    summary$inf.lb[bean.i] <- NA
+    summary$inf.ub[bean.i] <- NA
+
+  }
+
+
 }
 }
 
@@ -691,7 +806,7 @@ if(theme == 1) {
 
   if(is.null(point.o)) {point.o <- .2}
   if(is.null(bean.b.o)) {bean.b.o <- 1}
-  if(is.null(bean.f.o)) {bean.f.o <- .2}
+  if(is.null(bean.f.o)) {bean.f.o <- .3}
   if(is.null(inf.f.o)) {inf.f.o <- .8}
   if(is.null(inf.b.o)) {inf.b.o <- .8}
   if(is.null(avg.line.o)) {avg.line.o <- 1}
@@ -700,12 +815,14 @@ if(theme == 1) {
   if(is.null(bean.b.col)) {bean.b.col <- "black"}
   if(is.null(point.cex)) {point.cex <- .7}
   if(is.null(point.col)) {point.col <- "black"}
+  if(is.null(inf.b.col)) {inf.b.col <- gray(0)}
+  if(is.null(inf.f.col)) {inf.f.col <- "white"}
   if(is.null(bean.lwd)) {bean.lwd <- 2}
   if(is.null(avg.line.col)) {avg.line.col <- "black"}
 
 
   if(is.null(gl.col)) {gl.col <- "gray"}
-  if(is.null(gl.lwd)) {gl.lwd <- c(.5, 0)}
+  if(is.null(gl.lwd)) {gl.lwd <- c(.25, .5)}
 
   if(is.null(point.col)) {point.col <- "black"}
   if(is.null(point.bg)) {point.bg <- "white"}
@@ -724,6 +841,7 @@ if(theme == 2) {
   if(is.null(bean.f.o)) {bean.f.o <- 1}
   if(is.null(inf.f.o)) {inf.f.o <- .6}
   if(is.null(inf.b.o)) {inf.b.o <- .8}
+  if(is.null(inf.b.col)) {inf.b.col <- gray(0)}
   if(is.null(avg.line.o)) {avg.line.o <- 1}
   if(is.null(bar.f.o)) {bar.f.o <- 0}
   if(is.null(bar.b.o)) {bar.b.o <- 0}
@@ -737,7 +855,7 @@ if(theme == 2) {
   if(is.null(inf.disp)) {inf.disp <- "rect"}
 
   if(is.null(gl.col)) {gl.col <- "gray"}
-  if(is.null(gl.lwd)) {gl.lwd <- c(.5, 0)}
+  if(is.null(gl.lwd)) {gl.lwd <- c(.25, .5)}
 
 
 
@@ -757,7 +875,7 @@ if(theme == 3) {
 
   if(is.null(bean.b.col)) {bean.b.col <- "black"}
   if(is.null(inf.f.col)) {inf.f.col <- "white"}
-  if(is.null(inf.b.col)) {inf.b.col <-  "black"}
+  if(is.null(inf.b.col)) {inf.b.col <-  gray(0)}
   if(is.null(avg.line.col)) {avg.line.col <- "black"}
   if(is.null(point.col)) {point.col <- "black"}
   if(is.null(point.cex)) {point.cex <- .5}
@@ -766,7 +884,7 @@ if(theme == 3) {
   if(is.null(inf.disp)) {inf.disp <- "bean"}
 
   if(is.null(gl.col)) {gl.col <- "gray"}
-  if(is.null(gl.lwd)) {gl.lwd <- c(.5, 0)}
+  if(is.null(gl.lwd)) {gl.lwd <- c(.25, .5)}
 
 
 
@@ -785,7 +903,7 @@ if(theme == 4) {
   if(is.null(bar.b.o))  {bar.b.o <- 1}
 
   if(is.null(inf.f.col)) {inf.f.col <- "black"}
-  if(is.null(inf.b.col)) {inf.b.col <-  "black"}
+  if(is.null(inf.b.col)) {inf.b.col <-  gray(0)}
   if(is.null(avg.line.col)) {avg.line.col <- "black"}
   if(is.null(bar.f.col)) {bar.f.col <- "white"}
   if(is.null(bar.b.col)) {bar.b.col <- "black"}
@@ -796,7 +914,7 @@ if(theme == 4) {
   #  if(is.null(back.col)) {back.col <- gray(.97)}
 
   if(is.null(gl.col)) {gl.col <- "gray"}
-  if(is.null(gl.lwd)) {gl.lwd <- c(.5, 0)}
+  if(is.null(gl.lwd)) {gl.lwd <- c(.25, .5)}
 
   if(is.null(inf.disp)) {inf.disp <- "line"}
 
@@ -806,9 +924,8 @@ if(theme == 4) {
 # Inference lwd depends on inf.disp...
 if(is.null(inf.lwd)) {
 
-
-  if(inf.disp == "line") {inf.lwd <- 2}
-  if(inf.disp %in% c("rect", "bean")) {inf.lwd <- 1}
+  if(inf.disp == "line") {inf.lwd <- 3}
+  if(inf.disp %in% c("rect", "bean")) {inf.lwd <- 2}
 
 
 }
@@ -865,16 +982,16 @@ rownames(colors.df) <- 1:n.beans
 # If palette is in piratepal()...
 if(mean(pal %in% piratepal("names")) == 1) {
 
-colors.df$point.col <- rep(piratepal(palette = pal, length.out = n.cols), length.out = n.beans)
-colors.df$point.bg <- rep(piratepal(palette = pal, length.out = n.cols), length.out = n.beans)
-colors.df$bean.b.col <- rep(piratepal(palette = pal, length.out = n.cols), length.out = n.beans)
-colors.df$bean.f.col <- rep(piratepal(palette = pal, length.out = n.cols), length.out = n.beans)
-colors.df$inf.f.col <- rep(piratepal(palette = pal, length.out = n.cols), length.out = n.beans)
-colors.df$inf.b.col <- rep(piratepal(palette = pal, length.out = n.cols), length.out = n.beans)
-colors.df$avg.line.col <- rep(piratepal(palette = pal, length.out = n.cols), length.out = n.beans)
-colors.df$bar.f.col <- rep(piratepal(palette = pal, length.out = n.cols), length.out = n.beans)
-colors.df$bar.b.col <- rep(piratepal(palette = pal, length.out = n.cols), length.out = n.beans)
-colors.df$quant.col <- rep(piratepal(palette = pal, length.out = n.cols), length.out = n.beans)
+colors.df$point.col <- rep(piratepal(palette = pal, length.out = n.cols, mix.col = mix.col, mix.p = mix.p), length.out = n.beans)
+colors.df$point.bg <- rep(piratepal(palette = pal, length.out = n.cols, mix.col = mix.col, mix.p = mix.p), length.out = n.beans)
+colors.df$bean.b.col <- rep(piratepal(palette = pal, length.out = n.cols, mix.col = mix.col, mix.p = mix.p), length.out = n.beans)
+colors.df$bean.f.col <- rep(piratepal(palette = pal, length.out = n.cols, mix.col = mix.col, mix.p = mix.p), length.out = n.beans)
+colors.df$inf.f.col <- rep(piratepal(palette = pal, length.out = n.cols, mix.col = mix.col, mix.p = mix.p), length.out = n.beans)
+colors.df$inf.b.col <- rep(piratepal(palette = pal, length.out = n.cols, mix.col = mix.col, mix.p = mix.p), length.out = n.beans)
+colors.df$avg.line.col <- rep(piratepal(palette = pal, length.out = n.cols, mix.col = mix.col, mix.p = mix.p), length.out = n.beans)
+colors.df$bar.f.col <- rep(piratepal(palette = pal, length.out = n.cols, mix.col = mix.col, mix.p = mix.p), length.out = n.beans)
+colors.df$bar.b.col <- rep(piratepal(palette = pal, length.out = n.cols, mix.col = mix.col, mix.p = mix.p), length.out = n.beans)
+colors.df$quant.col <- rep(piratepal(palette = pal, length.out = n.cols, mix.col = mix.col, mix.p = mix.p), length.out = n.beans)
 
 }
 
@@ -1048,7 +1165,7 @@ if(is.null(ylim) == FALSE) {
 # Determine x and y labels
 
 if(subplot.n.iv == 1 & is.null(xlab)) {my.xlab <- iv.names[1]}
-if(subplot.n.iv == 1 & is.null(xlab) == F) {my.xlab <- xlab}
+if(subplot.n.iv == 1 & is.null(xlab) == FALSE) {my.xlab <- xlab}
 
 if(subplot.n.iv > 1) {my.xlab <- ""}
 
@@ -1072,6 +1189,7 @@ if(is.null(xlim)) {xlim <- c(min(bean.loc) - .5, max(bean.loc) + .5)}
        yaxt = "n",
        xlab = my.xlab,
        ylab = ylab,
+       cex.lab = cex.lab,
        main = main,
        yaxt = yaxt,
        bty = bty#,
@@ -1125,12 +1243,13 @@ rect(xleft = par("usr")[1],
 
 # GRIDLINES
 {
-if(is.null(gl.col) == F) {
+if(is.null(gl.col) == FALSE) {
 
-  if(is.null(gl.lwd)) {gl.lwd <- c(1, .5)}
+  if(is.null(gl.lwd)) {gl.lwd <- c(.5)}
   if(is.null(gl.lty)) {gl.lty <- 1}
+  if(is.null(gl)) {gl <- seq(min(y.levels), max(y.levels), length.out = length(y.levels))}
 
-  abline(h = seq(min(y.levels), max(y.levels), length.out = length(y.levels) * 2 - 1),
+  abline(h = gl,
          lwd = gl.lwd,
          col = gl.col,
          lty = gl.lty)
@@ -1473,16 +1592,16 @@ if(is.null(xaxt)) {
         side = 1,
         at = bean.mtx$x.loc,
         line = 1,
-        cex = cex.lab)
+        cex = cex.names)
 
 
   # IV 2 labels
 
   if(subplot.n.iv == 2) {
 
-    mtext(iv.names[2], side = 1, line = 2.5, at = par("usr")[1], adj = 1, cex = cex.lab)
+    mtext(iv.names[2], side = 1, line = 2.5, at = par("usr")[1], adj = 1, cex = cex.names)
 
-    mtext(iv.names[1], side = 1, line = 1, at = par("usr")[1], adj = 1, cex = cex.lab)
+    mtext(iv.names[1], side = 1, line = 1, at = par("usr")[1], adj = 1, cex = cex.names)
 
     text.loc <- (iv.lengths[1] + 1) / 2 * (2 *(1:iv.lengths[2]) - 1)
 
@@ -1510,7 +1629,7 @@ summary.df <- summary.df[,c(all.iv.names, setdiff(names(summary.df), all.iv.name
 
 output.ls <- list("summary" = summary.df,
                   "avg.line.fun" =  deparse(substitute(avg.line.fun)),
-                  "inf" = inf,
+                  "inf.method" = inf.method,
                   "inf.p" = inf.p)
 
 if(plot == FALSE) {